+++ routes/payment.js
@@ -5,6 +5,8 @@ const router = express.Router();
 require("dotenv").config();
 const paymentHistoryCollection = mongoClient.db("SoulMate").collection("paymentHistory");
 const orderCollection = mongoClient.db("SoulMate").collection("order");
+const usersCollection = mongoClient.db("SoulMate").collection("users");
+
 
 const stripe = require("stripe")(process.env.PAYMENT_KEY);
 const SSLCommerzPayment = require("sslcommerz-lts");
